@@ -31,114 +27,103 @@
     "license": "MIT",
     "sourceMap": true,
     "dependencies": {
-        "@babel/runtime": "^7.14.6",
-        "@fortawesome/fontawesome-svg-core": "^1.2.35",
-        "@fortawesome/free-brands-svg-icons": "^5.15.3",
-        "@fortawesome/free-regular-svg-icons": "^5.15.3",
-        "@fortawesome/free-solid-svg-icons": "^5.15.3",
-        "@fortawesome/react-fontawesome": "^0.1.14",
-        "@loadable/component": "^5.15.0",
-        "@loadable/server": "^5.14.0",
-        "@material-ui/core": "^4.12.3",
-        "@react-pdf/renderer": "^1.6.12",
-        "@tinymce/tinymce-react": "^3.8.3",
-        "axios": "^0.21.1",
-        "babel-polyfill": "^6.26.0",
-        "base64-arraybuffer": "^0.2.0",
-        "classnames": "^2.2.6",
-        "express": "4.16.0",
-        "formik": "^2.1.4",
-        "helmet": "3.13.0",
-        "history": "^4.7.2",
-        "moment": "^2.24.0",
-        "prop-types": "^15.7.2",
-        "react": "^16.14.0",
-        "react-bootstrap": "^1.6.1",
-        "react-datepicker": "^4.1.1",
-        "react-dom": "^16.14.0",
-        "react-google-charts": "^3.0.15",
-        "react-helmet": "5.2.0",
-        "react-image-lightbox": "^5.1.1",
-        "react-input-range": "^1.3.0",
+        "@babel/runtime": "^7.24.0",
+        "@emotion/react": "^11.11.0",
+        "@emotion/styled": "^11.11.0",
+        "@fortawesome/fontawesome-svg-core": "^6.5.0",
+        "@fortawesome/free-brands-svg-icons": "^6.5.0",
+        "@fortawesome/free-regular-svg-icons": "^6.5.0",
+        "@fortawesome/free-solid-svg-icons": "^6.5.0",
+        "@fortawesome/react-fontawesome": "^0.2.0",
+        "@mui/icons-material": "^5.15.0",
+        "@mui/material": "^5.15.0",
+        "@react-pdf/renderer": "^3.4.0",
+        "@tinymce/tinymce-react": "^4.0.2",
+        "axios": "^1.6.0",
+        "base64-arraybuffer": "^1.0.2",
+        "classnames": "^2.5.1",
+        "core-js": "^3.35.0",
+        "express": "^4.18.2",
+        "formik": "^2.4.5",
+        "helmet": "^7.1.0",
+        "history": "^5.3.0",
+        "moment": "^2.29.4",
+        "prop-types": "^15.8.1",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.9.0",
+        "react-datepicker": "^4.25.0",
+        "react-dom": "^18.2.0",
+        "react-google-charts": "^4.0.0",
+        "react-helmet-async": "^2.0.4",
         "react-js-pagination": "^3.0.3",
-        "react-loader-spinner": "3.1.2",
+        "react-loader-spinner": "^5.4.5",
         "react-month-picker-input": "^1.3.8",
-        "react-rating-stars-component": "2.2.0",
-        "react-redux": "5.1.1",
-        "react-responsive-modal": "^6.0.1",
-        "react-router": "^4.3.1",
-        "react-router-dom": "^4.3.1",
-        "react-router-redux": "^4.0.8",
-        "react-select-search": "^2.2.4",
-        "react-tooltip": "^4.2.6",
-        "react-universal-component": "3.0.3",
-        "redux": "4.0.1",
+        "react-rating-stars-component": "^2.2.0",
+        "react-redux": "^9.0.4",
+        "react-responsive-modal": "^7.1.0",
+        "react-router-dom": "^6.20.0",
+        "react-select-search": "^4.1.8",
+        "react-slider": "^2.0.6",
+        "react-tooltip": "^5.21.4",
+        "redux": "^5.0.0",
         "redux-immutable-state-invariant": "^2.1.0",
         "redux-logger": "^3.0.6",
-        "redux-saga": "^1.1.3",
-        "redux-thunk": "^2.3.0",
-        "reselect": "^4.0.0",
-        "socket.io": "^2.3.0",
-        "toastr": "2.1.4",
-        "uuid": "^8.3.2",
-        "webpack-flush-chunks": "2.0.1",
-        "winston": "2.4.1",
-        "yup": "^0.29.1"
+        "redux-saga": "^1.3.0",
+        "redux-thunk": "^3.1.0",
+        "reselect": "^5.0.1",
+        "socket.io": "^4.7.4",
+        "socket.io-client": "^4.7.4",
+        "toastr": "^2.1.4",
+        "uuid": "^9.0.1",
+        "winston": "^3.11.0",
+        "yet-another-react-lightbox": "^3.15.0",
+        "yup": "^1.4.0"
     },
     "devDependencies": {
-        "@babel/core": "^7.14.6",
-        "@babel/node": "^7.0.0",
-        "@babel/plugin-proposal-class-properties": "^7.0.0",
-        "@babel/plugin-proposal-decorators": "^7.0.0",
-        "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-        "@babel/plugin-transform-runtime": "^7.10.3",
-        "@babel/preset-env": "^7.0.0",
-        "@babel/preset-react": "^7.0.0",
-        "@loadable/babel-plugin": "^5.13.2",
-        "@loadable/webpack-plugin": "^5.14.0",
-        "autoprefixer": "9.1.5",
-        "babel-eslint": "^10.0.1",
-        "babel-jest": "^23.6.0",
-        "babel-loader": "^8.0.2",
-        "babel-plugin-universal-import": "^3.0.2",
-        "brotli-gzip-webpack-plugin": "^0.5.0",
-        "concurrently": "^5.3.0",
-        "copy-webpack-plugin": "4.6.0",
-        "core-js": "^3.1.4",
-        "cross-env": "^5.1.4",
-        "css-loader": "0.28.11",
-        "es": "^0.7.3",
-        "eslint": "5.6.1",
-        "eslint-plugin-react": "7.11.1",
-        "extract-css-chunks-webpack-plugin": "3.0.5",
-        "frameguard": "3.1.0",
-        "identity-obj-proxy": "3.0.0",
-        "immutable": "^4.0.0-rc.12",
-        "isomorphic-style-loader": "5.1.0",
-        "jest": "23.6.0",
+        "@babel/core": "^7.23.0",
+        "@babel/eslint-parser": "^7.24.0",
+        "@babel/node": "^7.23.0",
+        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+        "@babel/plugin-transform-class-properties": "^7.23.0",
+        "@babel/plugin-transform-private-methods": "^7.23.0",
+        "@babel/plugin-transform-runtime": "^7.23.0",
+        "@babel/preset-env": "^7.23.0",
+        "@babel/preset-react": "^7.23.0",
+        "ajv": "^8.12.0",
+        "ajv-keywords": "^5.1.0",
+        "autoprefixer": "^10.4.16",
+        "babel-jest": "^29.7.0",
+        "babel-loader": "^9.1.3",
+        "concurrently": "^8.2.2",
+        "copy-webpack-plugin": "^11.0.0",
+        "cross-env": "^7.0.3",
+        "css-loader": "^6.8.1",
+        "eslint": "^9.0.0",
+        "eslint-plugin-react": "^7.35.0",
+        "eslint-plugin-react-hooks": "^5.0.0",
+        "html-webpack-plugin": "^5.6.0",
+        "identity-obj-proxy": "^3.0.0",
+        "immutable": "^4.3.4",
+        "jest": "^29.7.0",
+        "jest-environment-jsdom": "^29.7.0",
         "lodash": "^4.17.21",
-        "net": "^1.0.2",
-        "node-sass": "^4.12.0",
-        "nodemon": "1.18.4",
+        "mini-css-extract-plugin": "^2.7.6",
+        "nodemon": "^3.0.2",
         "npm-run-all": "^4.1.5",
-        "postcss-import": "12.0.1",
-        "postcss-loader": "3.0.0",
-        "react-test-renderer": "16.5.2",
-        "rimraf": "^3.0.2",
-        "sass-loader": "7.1.0",
-        "stats-webpack-plugin": "0.6.2",
-        "style-loader": "0.20.3",
-        "stylus": "0.54.5",
-        "stylus-loader": "3.0.2",
-        "tls": "0.0.1",
+        "postcss": "^8.4.32",
+        "postcss-import": "^14.1.0",
+        "postcss-loader": "^7.3.3",
+        "react-test-renderer": "^18.2.0",
+        "rimraf": "^5.0.5",
+        "schema-utils": "^4.2.0",
+        "serve": "^14.2.1",
+        "style-loader": "^3.3.3",
         "url-loader": "^4.1.1",
-        "webpack": "^4.20.2",
-        "webpack-cli": "3.1.1",
-        "webpack-dev-middleware": "3.4.0",
-        "webpack-hot-middleware": "2.24.2",
-        "webpack-hot-server-middleware": "0.5.0",
-        "webpack-merge": "4.1.4",
-        "webpackbar": "2.6.3",
+        "webpack": "^5.89.0",
+        "webpack-cli": "^5.1.4",
+        "webpack-dev-server": "^4.15.1",
+        "webpack-merge": "^5.10.0",
+        "webpackbar": "^5.0.2",
         "xml-loader": "^1.2.1"
     }
 }
