@@ -1,25 +1,23 @@
 import React, { useEffect } from 'react';
-import { Route, Redirect } from 'react-router-dom';
-import AsyncComponent from '../shared/components/asyncComponent';
+import { Navigate, useLocation } from 'react-router-dom';
 
 const contextMenuHander = (e) => {
     e.preventDefault();
 };
 
-export const PrivateRoute = ({ component, ...rest }) => {
+export const PrivateRoute = ({ children }) => {
+    const location = useLocation();
+    
     useEffect(() => {
         window.addEventListener('contextmenu', contextMenuHander);
         return () => {
             window.removeEventListener('contextmenu', contextMenuHander);
         };
     }, []);
-    return (
-        <Route {...rest} component={props => (
-            localStorage.getItem('bloomkiteBusinessUser')
-                ? <AsyncComponent {...props} component={component} />
-                : <Redirect to={{ pathname: '/login', state: { from: props.location } }} />
-        )} />
-    );
+    
+    return localStorage.getItem('bloomkiteBusinessUser')
+        ? children
+        : <Navigate to="/login" state={{ from: location }} replace />;
 }
 
 export default PrivateRoute;
