@@ -1,9 +1,9 @@
-export default ({ appString, js, prodMeta, googleTracking, styles, helmet }) => `
+export default ({ appString, js, prodMeta, googleTracking, styles, helmet, preloadedState }) => `
     <!doctype html>
     <html lang="en">
         <head>
-            ${helmet.title.toString()}
-            ${helmet.meta.toString()}
+            ${helmet.title ? helmet.title.toString() : ''}
+            ${helmet.meta ? helmet.meta.toString() : ''}
             <meta charset="UTF-8"/>
             <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=5.0, minimum-scale=1.0"/>
             <meta http-equiv="X-UA-Compatible" content="ie=edge"/>
@@ -17,7 +17,6 @@ export default ({ appString, js, prodMeta, googleTracking, styles, helmet }) => 
             <link rel="preload" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" as="style" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
             <link rel="preload" href="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/latest/toastr.min.css" as="style">
             <link rel="preload" href="https://code.jquery.com/ui/1.11.3/themes/smoothness/jquery-ui.min.css" as="style" >
-            <link rel="preload" href="main.css" as="style" >
 
             <link rel="preload" href="https://ajax.googleapis.com/ajax/libs/jquery/3.4.0/jquery.min.js" as="script">
             <link rel="preload" href="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" as="script">
@@ -27,6 +26,7 @@ export default ({ appString, js, prodMeta, googleTracking, styles, helmet }) => 
             <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
             <link href="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/latest/toastr.min.css" rel="stylesheet"/>
             <link href="https://code.jquery.com/ui/1.11.3/themes/smoothness/jquery-ui.min.css" rel="stylesheet" />
+            <link rel="stylesheet" href="/app.client.css">
             ${styles}
             ${googleTracking}
         </head>
@@ -37,6 +37,10 @@ export default ({ appString, js, prodMeta, googleTracking, styles, helmet }) => 
             <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
             <script src="https://code.jquery.com/ui/1.11.3/jquery-ui.min.js"></script>
             <script src='https://checkout.razorpay.com/v1/checkout.js'></script>
+            <script>
+                window.__PRELOADED_STATE__ = ${preloadedState || '{}'};
+            </script>
+            <script  type="text/javascript" src="/app.client.js"></script>
             ${js}
         </body>
     </html>
