@@ -1,11 +1,11 @@
 import React from 'react';
-import ReactTooltip from 'react-tooltip';
+import { Tooltip } from 'react-tooltip';
 
 const CustomReactTooltip = (props) => {
     return (
-        <ReactTooltip className={props.className} id={props.id} type={props.type} effect={props.effect}>
+        <Tooltip className={props.className} id={props.id} variant={props.type} place={props.effect}>
             {props.children}
-        </ReactTooltip>
+        </Tooltip>
     );
 };
 
