@@ -1,13 +1,13 @@
 import React from 'react';
-import Loader from 'react-loader-spinner';
+import { Oval } from 'react-loader-spinner';
 
 const AppLoader = (props) => {
     const { loading } = props;
     return (
-        <div className="modal" className={loading ? 'modal show displayBlock' : 'displayHide'}>
+        <div className={loading ? 'modal show displayBlock' : 'displayHide'}>
             <div className="modal-dialog modal-loading">
                 <div className="">
-                    <Loader type="Bars" color="#FFF" />
+                    <Oval color="#FFF" height={40} width={40} />
                 </div>
             </div>
         </div>
