@@ -2,7 +2,7 @@ import React from 'react';
 
 const GettingStarted = () => (
     <section className="root">
-       
+        <h1>Getting Started</h1>
     </section>
 );
 
