@@ -7,7 +7,7 @@ import { numberMethod } from '../../../constants/commonRules';
 import Loader from '../../common/loader';
 import { insuranceSelector } from '../../../selectors/planning';
 import { calculateInsurance } from '../../../actions/planning';
-import { setTimeout } from 'timers';
+// setTimeout is available globally in both Node.js and browsers
 import Error from '../common/error';
 import { planningMessage } from '../../../constants/planningConstant';
 import { formatMoney, formatLoanAmount, Helper } from '../../../helpers/planningHelper';
