@@ -1,32 +1,27 @@
 import { createStore, compose, applyMiddleware } from 'redux';
 import reduxImmutableStateInvariant from 'redux-immutable-state-invariant';
-import thunk from 'redux-thunk';
+import { thunk } from 'redux-thunk';
 import { createLogger } from 'redux-logger';
 
-//import createHistory from 'history/createBrowserHistory';
-// 'routerMiddleware': the new way of storing route changes with redux middleware since rrV4.
-import { routerMiddleware } from 'react-router-redux';
+// Removed react-router-redux - using React Router v6
 import rootReducer from '../reducers';
 import createSagaMiddleware from 'redux-saga';
 import rootSaga from '../sagas';
 
 //export const history = createHistory();
 function configureStoreProd(initialState) {
-    const reactRouterMiddleware = routerMiddleware();
     const middlewares = [
         // Add other middleware on this line...
 
         // thunk middleware can also accept an extra argument to be passed to each thunk action
         // https://github.com/gaearon/redux-thunk#injecting-a-custom-argument
-        thunk,
-        reactRouterMiddleware
+        thunk
     ];
 
     return createStore(rootReducer, initialState, compose(applyMiddleware(...middlewares)));
 }
 
 function configureStoreDev(initialState) {
-    const reactRouterMiddleware = routerMiddleware();
     const sagaMiddleware = createSagaMiddleware();
     const middlewares = [
         // Add other middleware on this line...
@@ -36,8 +31,7 @@ function configureStoreDev(initialState) {
 
         // thunk middleware can also accept an extra argument to be passed to each thunk action
         // https://github.com/gaearon/redux-thunk#injecting-a-custom-argument
-        sagaMiddleware,
-        reactRouterMiddleware
+        sagaMiddleware
     ];
 
     const composeEnhancers = compose; // add support for Redux dev tools
